+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Changed
     - Truncate app setting values in ``remoteproject_sync.html`` (#1474)
     - JSON app setting value rendering in ``remoteproject_sync.html`` (#1472)
     - Change ``AppSettingAPI.compare_value()`` into public method (#1479)
+    - Refactor ``AppLinkContent`` (#1470, #1483)
 
 Fixed
 -----